backend/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==3.2.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-jwt==1.19.2
Pillow==9.5.0
pycparser==2.21
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
